rearrange code into smaller components
