+++ src/github/GithubExampleProps.ts
@@ -1,0 +1,6 @@
+interface GithubExampleProps {
+  setClientAccessToken: React.Dispatch<string>;
+  githubAccessToken: string;
+}
+
+export default GithubExampleProps;
